[CCM-244]: Standardize memory & cpu units for K8s data (#10723)
* Add kubernetes-java-client dependency

* Add K8sResourceStandardizer

* Have k8s utilization reporting use standard units

* Cleanup k8sResourceUtils

* Fix podinfo total resources

* Comments

* Fix assumptions about units in k8sResourceUtils

* Fix node-allocateable resources

* Change amount from String to long

* Convert instance units to ones used for price calculations

* Fix UTs

* Changes required on batch side

* Remove duplication

* Add UT coverage

+++ 78-batch-processing/src/main/java/io/harness/batch/processing/billing/timeseries/data/K8sGranularUtilizationData.java
@@ -10,8 +10,8 @@ public class K8sGranularUtilizationData {
   private String instanceId;
   private String instanceType;
   private String settingId;
-  private double cpu;
-  private double memory;
+  private long cpu;
+  private long memory;
   private long endTimestamp;
   private long startTimestamp;
 }
